# ENH: support for ND arithmetical operations for Function.