--- a/tests/unit/test_function.py
+++ b/tests/unit/test_function.py
@@ -283,7 +283,7 @@ def test_remove_outliers_iqr(x, y, expected_x, expected_y):
 def test_set_get_value_opt():
     """Test the set_value_opt and get_value_opt methods of the Function class."""
     func = Function(lambda x: x**2)
-    func.source = np.array([[1, 1], [2, 4], [3, 9], [4, 16], [5, 25]])
+    func.set_source(np.array([[1, 1], [2, 4], [3, 9], [4, 16], [5, 25]]))
     func.x_array = np.array([1, 2, 3, 4, 5])
     func.y_array = np.array([1, 4, 9, 16, 25])
     func.x_initial = 1