<!-- You are awesome! Your contribution to RocketPy is fundamental in our endeavour to create the next generation solution for rocketry trajectory simulation! -->

<!-- You may use this template to describe your Pull Request. But if you believe there is a better way to express yourself, don't hesitate! -->

## Pull request type
<!-- Remove unchecked box items. -->

- [X] Code changes (bugfix, features)

## Checklist
<!-- Remove irrelevant items to this PR. -->

- [X] Tests for the changes have been added (if needed)
- [X] Docs have been reviewed and added / updated
- [X] All tests (`pytest tests -m slow --runslow`) have passed locally
- [X] `CHANGELOG.md` has been updated (if relevant)

## Current behavior
<!-- Describe current behavior or link to an issue. -->

PR #785 introduced the possibility of 2D `Function` multiplication. The intent of this PR is to generalize this to the other arithmetical operations and to ND `Functions` for better scalability (e.g. so as not keep adding if blocks for every dimension, like 3D, 4D etc).

## New behavior
<!-- Describe changes introduced by this PR. -->

The generalization is done with the following changes:
- Use the `self._domain` and `self._image` for array manipulation, instead of unidimensional `self.x_array` and `self.y_array`;
- Multivariate lambda expressions are generated to allow callable operations:
    - The implementation is less trivial, requiring dynamically building a multivariate `lambda`;
    - This PR implementation worked very well on testing, but, if it does not hold review scrutinity due the use of `eval`, it can easily be converted into an `exception` and disallowing it.
    
Tests were introduced for both existing 2D operations and for the new generalized ones.

Overall the structure of the arithmetic methods of `Function` got a bit simpler and more regular. Another important notice, is that the current multivariable implementation of `mul` from #785 is non commutative when dealing with 1D*2D, which is fixed by this PR.

## Breaking change
<!-- If this introduces a breaking change, please describe the impact and migration path for existing applications below. Remove the unchecked box item. -->

- [ ] Yes
- [X] No