@phmbressan 2025-04-22
> Can you check execution speed of the code? With Monte Carlo sims especially
> 
> The operations seems different so I just want to make sure it is close to how it was before

### Performance Benchmarks

#### Function Composition (`f + g`) and Evaluation (`h(x)`)

| Test                         | Before                        | After                         |
|------------------------------|-------------------------------|-------------------------------|
| `f + g` (lambda)             | 16.2 μs ± 1.24 μs             | 17.5 μs ± 0.41 μs             |
| `h(3)` (lambda eval)         | 1.59 μs ± 0.013 μs            | 725 ns ± 5.47 ns              |
| `f + g` (array function)     | 110 μs ± 0.82 μs              | 102 μs ± 1.01 μs              |
| `h(2.5)` (array eval)        | 3.85 μs ± 0.045 μs            | 3.74 μs ± 0.068 μs            |

#### Monte Carlo Simulation

| Metric                   | Before         | After          |
|--------------------------|----------------|----------------|
| Avg Time per Iteration   | 1.337 s        | 1.276 s        |
| Total Simulations        | 250            | 250            |
| Total Wall Time          | 334.3 s        | 319.0 s        |

Overall the results are about the same speed, mostly within run to run variance. The exception is the Function lambda evaluation that is now consistently faster and I don't know why, still checking.